# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API Documentation",
    "description": "API Documentation",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:9999",
      "description": "Local server"
    }
  ],
  "tags": [
    {
      "name": "Auth",
      "description": "Auth"
    }
  ],
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "description": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/definitions/LoginUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logout",
        "description": "Logout",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register",
        "description": "Register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/definitions/RegisterUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Me",
        "description": "Me",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/car": {
      "get": {
        "tags": ["Car"],
        "summary": "Get all cars",
        "description": "Get all cars",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": ["Admin - Car"],
        "summary": "Add car",
        "description": "Add car",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/definitions/AddCarRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/car/{id}": {
      "get": {
        "tags": ["Car"],
        "summary": "Get car by id",
        "description": "Get car by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Car id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": ["Admin - Car"],
        "summary": "Update car",
        "description": "Update car",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Car id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/definitions/UpdateCarRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": ["Admin - Car"],
        "summary": "Delete car",
        "description": "Delete car",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Car id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/car/restore/{id}": {
      "put": {
        "tags": ["Admin - Car"],
        "summary": "Restore car",
        "description": "Restore car",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Car id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/car/force/{id}": {
      "delete": {
        "tags": ["Admin - Car"],
        "summary": "Force delete car",
        "description": "Force delete car",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Car id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/admin": {
      "get": {
        "tags": ["Super Admin - Admin"],
        "summary": "Get all admins",
        "description": "Get all admins",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": ["Super Admin - Admin"],
        "summary": "Add admin",
        "description": "Add admin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/definitions/RegisterUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/admin/{id}": {
      "get": {
        "tags": ["Super Admin - Admin"],
        "summary": "Get admin by id",
        "description": "Get admin by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Admin id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": ["Super Admin - Admin"],
        "summary": "Update admin",
        "description": "Update admin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Admin id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/definitions/UpdateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": ["Super Admin - Admin"],
        "summary": "Delete admin",
        "description": "Delete admin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Admin id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "definitions": {
      "LoginUserRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "user123"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        },
        "required": ["username", "password"]
      },
      "RegisterUserRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "User Name"
          },
          "username": {
            "type": "string",
            "example": "user123"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        },
        "required": ["username", "password", "name"]
      },
      "UpdateUserRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "User Name"
          },
          "username": {
            "type": "string",
            "example": "user123"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        }
      },
      "AddCarRequest": {
        "type": "object",
        "properties": {
          "plate": {
            "type": "string",
            "example": "AB123CD"
          },
          "manufacture": {
            "type": "string",
            "example": "Toyota"
          },
          "model": {
            "type": "string",
            "example": "Corolla"
          },
          "image": {
            "type": "string",
            "example": "http://example.com/car.jpg"
          },
          "rentPerDay": {
            "type": "number",
            "example": 100
          },
          "capacity": {
            "type": "number",
            "example": 5
          },
          "description": {
            "type": "string",
            "example": "A reliable family car"
          },
          "transmission": {
            "type": "string",
            "example": "Automatic"
          },
          "year": {
            "type": "number",
            "example": 2020
          }
        },
        "required": [
          "plate",
          "manufacture",
          "model",
          "image",
          "rentPerDay",
          "capacity",
          "description",
          "transmission",
          "year"
        ]
      },
      "UpdateCarRequest": {
        "type": "object",
        "properties": {
          "plate": {
            "type": "string",
            "example": "AB123CD"
          },
          "manufacture": {
            "type": "string",
            "example": "Toyota"
          },
          "model": {
            "type": "string",
            "example": "Corolla"
          },
          "image": {
            "type": "string",
            "example": "http://example.com/car.jpg"
          },
          "rentPerDay": {
            "type": "number",
            "example": 100
          },
          "capacity": {
            "type": "number",
            "example": 5
          },
          "description": {
            "type": "string",
            "example": "A reliable family car"
          },
          "transmission": {
            "type": "string",
            "example": "Automatic"
          },
          "year": {
            "type": "number",
            "example": 2020
          }
        }
      }
    }
  }
}